+++ cpp/dolfinx/mesh/graphbuild.cpp
@@ -522,7 +522,7 @@ std::tuple<graph::AdjacencyList<std::int32_t>, std::vector<std::int64_t>,
 mesh::build_local_dual_graph(
     std::span<const CellType> celltypes,
     const std::vector<std::span<const std::int64_t>>& cells,
-    std::optional<std::uint16_t> max_facet_to_cell_links)
+    std::optional<std::uint32_t> max_facet_to_cell_links)
 {
   spdlog::info("Build local part of mesh dual graph (mixed)");
   common::Timer timer("Compute local part of mesh dual graph (mixed)");
@@ -729,7 +729,7 @@ mesh::build_local_dual_graph(
 graph::AdjacencyList<std::int64_t>
 mesh::build_dual_graph(MPI_Comm comm, std::span<const CellType> celltypes,
                        const std::vector<std::span<const std::int64_t>>& cells,
-                       std::optional<std::uint16_t> max_facet_to_cell_links)
+                       std::optional<std::uint32_t> max_facet_to_cell_links)
 {
   spdlog::info("Building mesh dual graph");
